refactored to keep the settings after crashing
